core/requirements.txt
# sentinel - dependencias python
# c1q_ (M-Society team)

# framework web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# async http
httpx==0.27.0
aiohttp==3.9.3

# scraping
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.2
selenium-wire==5.1.0

# network
scapy==2.5.0
python-nmap==0.7.1
dnspython==2.6.1
python-whois==0.9.4

# procesamiento
pandas==2.2.0
numpy==1.26.4
orjson==3.9.13

# nlp
spacy==3.7.4
sentence-transformers==2.3.1

# colas
celery==5.3.6
redis==5.0.1

# orm y db
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# elasticsearch
elasticsearch[async]==8.12.1

# neo4j
neo4j==5.17.0

# validacion
pydantic==2.6.1
pydantic-settings==2.1.0

# auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# logging
structlog==24.1.0

# reportes
weasyprint==61.2
jinja2==3.1.3

# metadata
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0

# telefonia
phonenumbers==8.13.29

# tor
stem==1.8.2
PySocks==1.7.1

# utilidades
pyyaml==6.0.1
python-dotenv==1.0.1
tenacity==8.2.3
cryptography==42.0.2

# testing
pytest==8.0.1
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx-sse==0.4.0

# grpc
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3

# stix
stix2==3.0.1

# prometheus
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
